added search, palette btn design
made the palette button background designs consistent with the rest of the site, now it is the acutal buttons themselves left to be redesigned. as well as the palettes themselves.
Added search functionality to the color tree and tested it sufficiently, it works just fine.. Buttons are left to be designed here as well